set many to many sequalize

+++ src/models/companydetail.js
@@ -29,7 +29,7 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.TEXT
     },
     phone_number: {
-      type: DataTypes.INTEGER
+      type: DataTypes.BIGINT
     },
     status_disband: {
       type: DataTypes.BOOLEAN,
